requirements.txt
# Core dependencies
beautifulsoup4==4.12.2
feedparser==6.0.10
requests==2.31.0
openai==1.3.0

# TTS options (chọn 1)
gTTS==2.3.2              # Google TTS (free)
# elevenlabs==0.2.0     # ElevenLabs (uncomment nếu dùng)

# GitHub integration
PyGithub==2.1.1

# Configuration & utilities
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2

# Optional: audio processing
# pydub==0.25.1         # Nếu cần mix/nhạc nền